Move checks to the admin app
