+++ core/port/services.go
@@ -1,9 +1,6 @@
 package port
 
 import (
-	"context"
-
-	"github.com/vertex-center/vertex/core/types"
 	"github.com/vertex-center/vertex/core/types/app"
 )
 
@@ -14,9 +11,5 @@ type (
 
 	DebugService interface {
 		HardReset()
-	}
-
-	ChecksService interface {
-		CheckAll(ctx context.Context) <-chan types.CheckResponse
 	}
 )
